Add Snacks recycler & soil storage to Itinerant
The Snacks mod adds a SoilRecycler module and Soil RESOURCE block to
several stock hab parts, including the Hitchhiker.  SSPXR does the same
with its hab parts.  The Itinerant is basically a half-capacity version
of the Hitchhiker, so it ought have the same features.

This is basically just a copy of the relevant portions of
SSPXR-Snacks.cfg from the SSPXR mod.  I've copied the recycler's
localized strings in the three languages for which SSPXR has them:
English, Russian, and Chinese.  (Other languages lack translations for
the Snacks recycler strings, both here and in SSPXR.)